# gateway/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "uRouter Gateway API",
    "version": "0.1.0",
    "description": "P0 compatibility baseline. Schemas are versioned independently from the binary."
  },
  "servers": [{ "url": "http://127.0.0.1:8787" }],
  "paths": {
...
    "/v1/chat/completions": { "post": { "operationId": "chatCompletions", "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }], "requestBody": { "$ref": "#/components/requestBodies/ChatRequest" }, "responses": { "200": { "description": "OpenAI-compatible completion or stream", "headers": { "x-urouter-request-id": { "$ref": "#/components/headers/RequestId" }, "x-urouter-decision-id": { "$ref": "#/components/headers/DecisionId" } } }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } } },
...
  },
  "components": {
    "securitySchemes": { "managementBearer": { "type": "http", "scheme": "bearer" } },
    "headers": { "RequestId": { "description": "Opaque logical Gateway request identifier", "schema": { "type": "string" } }, "DecisionId": { "description": "Opaque immutable routing decision identifier when a decision exists", "schema": { "type": "string" } } },
    "parameters": { "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }, "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": false, "description": "Tenant-scoped logical operation key. Reuse with a different canonical request returns 409; response replay is not currently provided.", "schema": { "type": "string", "minLength": 1, "maxLength": 256, "pattern": "^[!-~]+$" } } },
    "requestBodies": { "ChatRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatRequest" } } } } },
    "responses": { "Error": { "description": "uRouter error", "headers": { "x-urouter-request-id": { "$ref": "#/components/headers/RequestId" }, "x-urouter-decision-id": { "$ref": "#/components/headers/DecisionId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } } },
    "schemas": {
      "ChatRequest": { "type": "object", "required": ["model", "messages"], "properties": { "model": { "type": "string" }, "messages": { "type": "array", "items": { "type": "object" } }, "stream": { "type": "boolean" }, "tools": { "type": "array", "items": { "type": "object" } }, "urouter": { "type": "object" } }, "additionalProperties": true },
...
      "ErrorEnvelope": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message", "type"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "type": { "const": "urouter_error" } } } } }
    }
  }
}